requirements.txt
Flask
requests
Pillow
azure-storage-blob
google-api-python-client
google-auth-oauthlib
gunicorn

